mpegtsbase: Prevent stack gardening by using the right type
stream_type is stored as guint inside the GstStructure but was retreived
using valist with a pointer to guint16. This would cause stack gardening
when code is compiled without optimisation (e.g. in -O0 the compiler wont
pad the stack to optimise out required mask).

https://bugzilla.gnome.org/show_bug.cgi?id=655540

+++ gst/mpegtsdemux/mpegtsbase.c
@@ -615,7 +615,7 @@ mpegts_base_activate_program (MpegTSBase * base, MpegTSBaseProgram * program,
   guint i, nbstreams;
   guint pcr_pid;
   guint pid;
-  guint16 stream_type;
+  guint stream_type;
   GstStructure *stream;
   const GValue *new_streams;
   const GValue *value;
